// fediverse-server/src/types/user.ts
export interface User {
  fediverseId: string;
  username: string;
  displayName: string;
  summary?: string;
  profilePicUrl: string;
  inbox?: string;
  outbox?: string;
  followers?: string;
  following?: string;
  publicKey: string;
  encryptedPrivateKey?: string;
  domain?: string;
  isFederated?: boolean;
  followersCount?: number;
  followingCount?: number;
}